+++ EKF/include/EKFFixedLag.h
@@ -11,12 +11,12 @@
  * Structure to hold historical filter state for smoothing.
  */
 struct FilterState {
-    double t;                // Time t_k
-    Eigen::VectorXd x_pred;  // x_{k|k-1}
-    Eigen::MatrixXd P_pred;  // P_{k|k-1}
-    Eigen::VectorXd x_upd;   // x_{k|k} (filtered)
-    Eigen::MatrixXd P_upd;   // P_{k|k} (filtered)
-    Eigen::MatrixXd F;       // F_k (Jacobian used to predict x_{k|k-1} from x_{k-1|k-1})
+    float t;                 // Time t_k
+    Eigen::VectorXf x_pred;  // x_{k|k-1}
+    Eigen::MatrixXf P_pred;  // P_{k|k-1}
+    Eigen::VectorXf x_upd;   // x_{k|k} (filtered)
+    Eigen::MatrixXf P_upd;   // P_{k|k} (filtered)
+    Eigen::MatrixXf F;       // F_k (Jacobian used to predict x_{k|k-1} from x_{k-1|k-1})
 };
 
 /**
@@ -30,7 +30,7 @@ struct FilterState {
  */
 class EKFFixedLag {
 public:
-    EKFFixedLag(SystemModel* model, const Eigen::VectorXd& x0, const Eigen::MatrixXd& P0, int lag_L);
+    EKFFixedLag(SystemModel* model, const Eigen::VectorXf& x0, const Eigen::MatrixXf& P0, int lag_L);
 
     /**
      * Perform one time step.
@@ -38,19 +38,19 @@ public:
      * @param u_k Control input at current time
      * @param t_k Current time
      */
-    void step(const Eigen::VectorXd& y_k, const Eigen::VectorXd& u_k, double t_k);
+    void step(const Eigen::VectorXf& y_k, const Eigen::VectorXf& u_k, float t_k);
 
     /**
      * Get the filtered state at current time k (x_{k|k}).
      */
-    std::pair<Eigen::VectorXd, Eigen::MatrixXd> getFilteredState() const;
+    std::pair<Eigen::VectorXf, Eigen::MatrixXf> getFilteredState() const;
 
     /**
      * Get the smoothed state at current time k (x_{k|k}).
      * This is the result of the backward pass starting at k, so it's equal to filtered at k,
      * but smoothed for previous steps.
      */
-    std::pair<Eigen::VectorXd, Eigen::MatrixXd> getSmoothedState(int lag) const;
+    std::pair<Eigen::VectorXf, Eigen::MatrixXf> getSmoothedState(int lag) const;
 
 private:
     SystemModel* model_;
@@ -62,8 +62,8 @@ private:
 
     // Smoothed estimates cache (recomputed each step)
     // Organized same as buffer_: Index 0 corresponds to buffer_[0] (k-L)
-    std::vector<Eigen::VectorXd> x_smooth_;
-    std::vector<Eigen::MatrixXd> P_smooth_;
+    std::vector<Eigen::VectorXf> x_smooth_;
+    std::vector<Eigen::MatrixXf> P_smooth_;
 };
 
 #endif // EKF_FIXED_LAG_H
