+++ PKF/src/example_main.cpp
@@ -24,18 +24,18 @@ int main() {
 
     // Initialize True State
     Lorenz63Model::State x_true;
-    x_true << 1.0, 1.0, 1.0; // Start slightly off-center
+    x_true << 1.0f, 1.0f, 1.0f; // Start slightly off-center
 
     // Initialize Filter
     // Prior: Gaussian around initial truth with some spread
     auto prior_sampler = [&](std::mt19937_64& r) -> Lorenz63Model::State {
-        return Noise::gaussian_sample<3>(x_true, 5.0 * Lorenz63Model::StateMat::Identity(), r);
+        return Noise::gaussian_sample<3>(x_true, 5.0f * Lorenz63Model::StateMat::Identity(), r);
     };
     pf.initialize(prior_sampler);
 
     // Data Storage
     struct DataPoint {
-        double t;
+        float t;
         Lorenz63Model::State x_true;
         Lorenz63Model::Observation y;
         Lorenz63Model::State x_est_filt;
@@ -51,7 +51,7 @@ int main() {
     std::cout << "Particles: " << NUM_PARTICLES << ", Lag: " << LAG << std::endl;
 
     for (int k = 1; k <= STEPS; ++k) {
-        double t = k * Lorenz63Model::DT;
+        float t = k * Lorenz63Model::DT;
 
         // 1. Simulate Truth
         x_true = model.propagate(x_true, t, u) + model.sample_process_noise(t, rng);
@@ -63,26 +63,6 @@ int main() {
         pf.step(y, t, u);
 
         // 4. Store Results
-        // Note: Smoothed estimate lags by LAG.
-        // At step k, get_smoothed_mean() returns estimate for k-LAG (or oldest available).
-        // To align for plotting, we usually align by time index.
-        // But here we'll just dump what we have.
-        // We will store: Current Truth, Current Filtered, Current Smoothed (which corresponds to k-LAG effectively if buffer full)
-        // Actually, to plot correctly, we need to map the smoothed estimate to its correct time.
-        // But for simple CSV dumping:
-        // Let's dump:
-        // t, x_true_x, x_true_y, x_true_z, filt_x, filt_y, filt_z, smooth_x, smooth_y, smooth_z
-        // AND we should note that smooth_x at row k is actually the smoothed estimate for time k-LAG.
-        // OR we can run the smoother offline? No, it's online fixed-lag.
-
-        // Wait, the standard way to evaluate fixed lag smoother is to compare:
-        // x_{k-L} (True state L steps ago) vs x_{k-L|k} (Smoothed estimate)
-
-        // To simplify plotting: we will just log everything.
-        // The plotting script can handle the lag shift if we provide the time stamp for the smoothed estimate.
-        // But `get_smoothed_mean` doesn't return the time.
-        // However, we know it is k-L (clamped at 0).
-
         results.push_back({t, x_true, y, pf.get_filtered_mean(), pf.get_smoothed_mean()});
     }
