+++ PKF/tests/test_particle.cpp
@@ -12,38 +12,38 @@
 // y_k = x_k + v_k
 class TestModel1D : public PKF::StateSpaceModel<1, 1> {
 public:
-    State propagate(const State& x_prev, double t_k, const Eigen::Ref<const State>& u_k) const override {
+    State propagate(const State& x_prev, float t_k, const Eigen::Ref<const State>& u_k) const override {
         (void)t_k;
         return x_prev + u_k;
     }
 
-    Observation observe(const State& x_k, double t_k) const override {
+    Observation observe(const State& x_k, float t_k) const override {
         (void)t_k;
         return x_k;
     }
 
-    State sample_process_noise(double t_k, std::mt19937_64& rng) const override {
+    State sample_process_noise(float t_k, std::mt19937_64& rng) const override {
         (void)t_k;
-        std::normal_distribution<double> d(0.0, 0.1);
+        std::normal_distribution<float> d(0.0f, 0.1f);
         return State::Constant(d(rng));
     }
 
-    Observation sample_observation_noise(double t_k, std::mt19937_64& rng) const override {
+    Observation sample_observation_noise(float t_k, std::mt19937_64& rng) const override {
         (void)t_k;
-        std::normal_distribution<double> d(0.0, 0.5);
+        std::normal_distribution<float> d(0.0f, 0.5f);
         return Observation::Constant(d(rng));
     }
 
-    double observation_loglik(const Observation& y_k, const State& x_k, double t_k) const override {
+    float observation_loglik(const Observation& y_k, const State& x_k, float t_k) const override {
         (void)t_k;
-        double diff = y_k(0) - x_k(0);
-        return -0.5 * (std::log(2 * M_PI * 0.25) + diff * diff / 0.25);
+        float diff = y_k(0) - x_k(0);
+        return -0.5f * (std::log(2.0f * static_cast<float>(M_PI) * 0.25f) + diff * diff / 0.25f);
     }
 };
 
 void test_resampling() {
     std::cout << "Testing Resampling..." << std::endl;
-    std::vector<double> weights = {0.1, 0.2, 0.3, 0.4};
+    std::vector<float> weights = {0.1f, 0.2f, 0.3f, 0.4f};
     std::mt19937_64 rng(42);
 
     auto parents = PKF::Resampling::systematic(weights, rng);
@@ -62,22 +62,22 @@ void test_particle_filter() {
 
     // Initialize
     pf.initialize([](std::mt19937_64& r) {
-        std::normal_distribution<double> d(0.0, 1.0);
-        return Eigen::Matrix<double, 1, 1>::Constant(d(r));
+        std::normal_distribution<float> d(0.0f, 1.0f);
+        return Eigen::Matrix<float, 1, 1>::Constant(d(r));
     });
 
     // Step
-    Eigen::Matrix<double, 1, 1> u; u << 1.0;
-    Eigen::Matrix<double, 1, 1> y; y << 1.0;
+    Eigen::Matrix<float, 1, 1> u; u << 1.0f;
+    Eigen::Matrix<float, 1, 1> y; y << 1.0f;
 
-    pf.step(y, 1.0, u);
+    pf.step(y, 1.0f, u);
 
     auto mean = pf.get_mean();
     std::cout << "Filtered Mean: " << mean(0) << std::endl;
 
     // Basic sanity check: mean should be somewhere near 1.0 (prior 0 + u 1)
     // and dragged towards y=1.
-    assert(std::abs(mean(0) - 1.0) < 1.0);
+    assert(std::abs(mean(0) - 1.0f) < 1.0f);
 
     std::cout << "Particle Filter tests passed." << std::endl;
 }
